fix(app): open transaction rule pills in context
