--- packages/app/src/app/(tabs)/settings/rules-page.selector.ts
@@ -1,9 +1,0 @@
-export const RulesPageSelector = {
-    Page: 'RulesPage.Page',
-    EmptyState: 'RulesPage.EmptyState',
-    CreateButton: 'RulesPage.CreateButton',
-    RuleCard: (index: number) => `RulesPage.RuleCard.${index}` as const,
-    RuleCardEnabledSwitch: (index: number) => `RulesPage.RuleCard.${index}.EnabledSwitch` as const,
-    RuleCardConditions: (index: number) => `RulesPage.RuleCard.${index}.Conditions` as const,
-    RuleCardActions: (index: number) => `RulesPage.RuleCard.${index}.Actions` as const
-} as const;
