+++ packages/app/src/app/(tabs)/settings/rules.tsx
@@ -1,124 +1,14 @@
-import { RuleWithActionsRelationsEntityInterface, UserIconNameEnum } from '@budgie/contracts';
-import { LegendList } from '@legendapp/list';
-import { useLingui } from '@lingui/react/macro';
-import { NotificationFeedbackType } from 'expo-haptics/src/Haptics.types';
-import { ReactElement, useState } from 'react';
-import { View } from 'react-native';
-import Toast from 'react-native-toast-message';
+import { useLocalSearchParams } from 'expo-router';
 
-import { getErrorMessage, isDefined, isNotEmptyArray } from '@rnw-community/shared';
-
-import { DeletableRow } from '../../../@generic/component/deletable-row/deletable-row';
-import { MenuSpacer } from '../../../@generic/component/menu-spacer/menu-spacer';
-import { Page } from '../../../@generic/component/page/page';
-import { PageHeader } from '../../../@generic/component/page-header/page-header';
-import { SearchablePageEmptyState } from '../../../@generic/component/searchagle-page-empty-state/searchagle-page-empty-state';
-import {
-    LEGEND_LIST_CONTENT_GAP,
-    LEGEND_LIST_ESTIMATED_ITEM_SIZE,
-    LEGEND_LIST_HEADER_HEIGHT,
-    LEGEND_LIST_STYLE
-} from '../../../@generic/constant/legend-list.constant';
-import { useCreateAction } from '../../../@generic/hook/use-create-action.hook';
-import { useVibration } from '../../../@generic/hook/use-vibration.hook';
 import { goBackOrReplace } from '../../../@generic/utils/go-back-or-replace.util';
-import { RuleCard } from '../../../rule/components/rule-card/rule-card';
-import { useRuleFormModal } from '../../../rule/context/rule-form-modal.context';
-import { useGetAllRulesQuery } from '../../../rule/query/use-get-all-rules.query';
-import { ruleService } from '../../../rule/service/rule.service';
+import { RulesListPage } from '../../../rule/components/rules-list-page/rules-list-page';
+import { parseRuleIdsRouteParam } from '../../../rule/utils/parse-rule-ids-route-param.util';
 
-import { RulesPageSelector } from './rules-page.selector';
-
-const CONTENT_CONTAINER_STYLE = { gap: LEGEND_LIST_CONTENT_GAP };
-const HEADER_SPACER_STYLE = { height: LEGEND_LIST_HEADER_HEIGHT };
-
-const listHeader = <View style={HEADER_SPACER_STYLE} />;
-const listFooter = <MenuSpacer />;
 const handleGoBack = () => void goBackOrReplace('/settings');
-const getKeyExtractor = (item: Pick<RuleWithActionsRelationsEntityInterface, 'id'>) => item.id.toString();
 
 export default function RulesPage() {
-    const { t } = useLingui();
-    const [rulesRefreshKey, setRulesRefreshKey] = useState(0);
-    const { rules } = useGetAllRulesQuery(rulesRefreshKey);
-    const [notify] = useVibration();
-    const { openRuleForm } = useRuleFormModal();
+    const routeParams = useLocalSearchParams();
+    const matchingRuleIds = parseRuleIdsRouteParam(routeParams['ruleIds']);
 
-    const refreshRules = () => {
-        setRulesRefreshKey(value => value + 1);
-    };
-
-    const handleRuleFormResult = (result: Awaited<ReturnType<typeof openRuleForm>>) => {
-        if (isDefined(result)) {
-            refreshRules();
-        }
-    };
-
-    const handleDeleteRule = async (id: number) => {
-        await ruleService.archiveById(id);
-        refreshRules();
-        notify(NotificationFeedbackType.Success);
-    };
-
-    const handleToggleRule = async (rule: Pick<RuleWithActionsRelationsEntityInterface, 'id'>, enabled: boolean) => {
-        try {
-            await ruleService.toggleEnabled(rule.id, enabled);
-            refreshRules();
-        } catch (error: unknown) {
-            Toast.show({ type: 'error', text1: t`Could not update rule`, text2: getErrorMessage(error) });
-        }
-    };
-
-    const handleOpenRule = (rule: Pick<RuleWithActionsRelationsEntityInterface, 'id'>) =>
-        void openRuleForm({ ruleId: rule.id }).then(handleRuleFormResult);
-
-    const handleCreateRule = () => void openRuleForm().then(handleRuleFormResult);
-
-    useCreateAction({
-        icon: UserIconNameEnum.Zap,
-        label: t`Rule`,
-        variant: 'primary',
-        onPress: handleCreateRule,
-        testID: RulesPageSelector.CreateButton
-    });
-
-    const renderItem = ({ item, index }: { item: RuleWithActionsRelationsEntityInterface; index: number }): ReactElement => (
-        <DeletableRow id={item.id} onDelete={handleDeleteRule}>
-            <RuleCard
-                testID={RulesPageSelector.RuleCard(index)}
-                switchTestID={RulesPageSelector.RuleCardEnabledSwitch(index)}
-                conditionsTestID={RulesPageSelector.RuleCardConditions(index)}
-                actionsTestID={RulesPageSelector.RuleCardActions(index)}
-                onOpen={handleOpenRule}
-                onToggle={handleToggleRule}
-                rule={item}
-            />
-        </DeletableRow>
-    );
-
-    return (
-        <Page testID={RulesPageSelector.Page} header={<PageHeader onGoBack={handleGoBack} title={t`Rules`} />} withBlur>
-            {isNotEmptyArray(rules) ? (
-                <LegendList
-                    style={LEGEND_LIST_STYLE}
-                    contentContainerStyle={CONTENT_CONTAINER_STYLE}
-                    ListHeaderComponent={listHeader}
-                    data={rules}
-                    renderItem={renderItem}
-                    keyExtractor={getKeyExtractor}
-                    estimatedItemSize={LEGEND_LIST_ESTIMATED_ITEM_SIZE}
-                    recycleItems
-                    showsVerticalScrollIndicator={false}
-                    ListFooterComponent={listFooter}
-                />
-            ) : (
-                <SearchablePageEmptyState
-                    testID={RulesPageSelector.EmptyState}
-                    title={t`No Rules Yet`}
-                    icon={UserIconNameEnum.Zap}
-                    description={t`Create rules to automatically categorize and tag your bank transactions`}
-                />
-            )}
-        </Page>
-    );
+    return <RulesListPage matchingRuleIds={matchingRuleIds} onGoBack={handleGoBack} />;
 }
