+++ packages/app/src/rule/interface/rules-list-page-props.interface.ts
@@ -1,0 +1,4 @@
+export interface RulesListPagePropsInterface {
+    readonly matchingRuleIds?: readonly number[];
+    readonly onGoBack: () => void;
+}
