+++ tests/app-tests/flows/26.rules-application.flow.yaml
@@ -1032,7 +1032,7 @@ env:
 - assertVisible:
     id: 'RulesPage.EmptyState'
 
-# --- Matching Rules Pill (import → verify pill on transaction → navigate to rules) ---
+# --- Matching Rules Pill (import → verify pill on transaction → open matching rule) ---
 
 - tapOn:
     id: 'ActionButton'
@@ -1117,7 +1117,5 @@ env:
 - waitForAnimationToEnd
 
 - extendedWaitUntil:
-    visible:
-      id: 'RulesPage.Page'
-- assertVisible:
-    id: 'RulesPage.Page'
+    visible: 'Edit Rule'
+- assertVisible: 'Edit Rule'
