+++ tests/app-tests/flows/27.matching-rules-navigation.flow.yaml
@@ -1,0 +1,204 @@
+appId: ${APP_ID}
+env:
+  ACCOUNT_NAME: Rules Test Account
+  FIRST_RULE_MATCH: E2E Pill Navigation
+  SECOND_RULE_MATCH: Pill Navigation
+  TAG_NAME: Rules Test Tag
+  TRANSACTION_CATEGORY_NAME: Rules Test Category
+---
+- runFlow:
+    file: setup/setup-rules-test.flow.yaml
+
+- runFlow:
+    file: subflows/navigate-to-rules.flow.yaml
+
+- tapOn:
+    id: 'ActionButton'
+- extendedWaitUntil:
+    visible:
+      id: 'RulesPage.CreateButton'
+- waitForAnimationToEnd
+- tapOn:
+    id: 'RulesPage.CreateButton'
+- extendedWaitUntil:
+    visible: 'Create Rule'
+
+- tapOn:
+    id: 'RuleForm.Condition.0.FieldSelector'
+- waitForAnimationToEnd
+- tapOn: 'Comment'
+- waitForAnimationToEnd
+- tapOn:
+    id: 'RuleForm.Condition.0.ValueInput'
+- inputText: ${FIRST_RULE_MATCH}
+- tapOn: return
+
+- tapOn:
+    id: 'RuleForm.Action.0.TypeSelector'
+- tapOn:
+    id: 'RuleForm.SelectorCard.SET_CATEGORY'
+- waitForAnimationToEnd
+
+- scrollUntilVisible:
+    element:
+      id: 'RuleForm.Action.0.CategorySelector'
+    direction: DOWN
+    centerElement: true
+
+- tapOn:
+    id: 'RuleForm.Action.0.CategorySelector'
+- runFlow:
+    file: subflows/shared/select-category-by-search.flow.yaml
+    env:
+      CATEGORY_NAME: Groceries
+
+- tapOn:
+    id: 'RuleForm.Footer.SubmitButton'
+- runFlow:
+    when:
+      visible: 'Apply to existing transactions?'
+    commands:
+      - tapOn: 'Skip'
+- waitForAnimationToEnd
+
+- assertVisible:
+    id: 'RulesPage.RuleCard.0'
+
+- runFlow:
+    file: subflows/transactions/create-expense.flow.yaml
+    env:
+      AMOUNT: 25
+      ACCOUNT_NAME: ${ACCOUNT_NAME}
+      CATEGORY_NAME: ${TRANSACTION_CATEGORY_NAME}
+      COMMENT: ${FIRST_RULE_MATCH}
+
+- extendedWaitUntil:
+    visible:
+      id: 'TransactionCard.Label.E2E_Pill_Navigation'
+    timeout: 10000
+- tapOn:
+    id: 'TransactionCard.Label.E2E_Pill_Navigation'
+- waitForAnimationToEnd
+
+- extendedWaitUntil:
+    visible:
+      id: 'MatchingRules.Pill'
+    timeout: 10000
+- tapOn:
+    id: 'MatchingRules.Pill'
+- waitForAnimationToEnd
+
+- extendedWaitUntil:
+    visible: 'Edit Rule'
+- assertVisible: 'Edit Rule'
+
+- tapOn:
+    id: 'PageHeader.BackButton'
+- waitForAnimationToEnd
+- assertVisible: 'Edit Expense'
+
+- tapOn:
+    id: 'PageHeader.BackButton'
+- waitForAnimationToEnd
+
+- runFlow:
+    file: subflows/navigate-to-rules.flow.yaml
+
+- tapOn:
+    id: 'ActionButton'
+- extendedWaitUntil:
+    visible:
+      id: 'RulesPage.CreateButton'
+- waitForAnimationToEnd
+- tapOn:
+    id: 'RulesPage.CreateButton'
+- extendedWaitUntil:
+    visible: 'Create Rule'
+
+- tapOn:
+    id: 'RuleForm.Condition.0.FieldSelector'
+- waitForAnimationToEnd
+- tapOn: 'Comment'
+- waitForAnimationToEnd
+- tapOn:
+    id: 'RuleForm.Condition.0.ValueInput'
+- inputText: ${SECOND_RULE_MATCH}
+- tapOn: return
+
+- tapOn:
+    id: 'RuleForm.Action.0.TypeSelector'
+- tapOn:
+    id: 'RuleForm.SelectorCard.ADD_TAG'
+- waitForAnimationToEnd
+
+- scrollUntilVisible:
+    element:
+      id: 'RuleForm.Action.0.TagSelector'
+    direction: DOWN
+    centerElement: true
+
+- tapOn:
+    id: 'RuleForm.Action.0.TagSelector'
+- waitForAnimationToEnd
+- tapOn:
+    text: ${TAG_NAME}
+
+- tapOn:
+    id: 'RuleForm.Footer.SubmitButton'
+- runFlow:
+    when:
+      visible: 'Apply to existing transactions?'
+    commands:
+      - tapOn: 'Skip'
+- waitForAnimationToEnd
+
+- assertVisible:
+    id: 'RulesPage.RuleCard.0'
+- assertVisible:
+    id: 'RulesPage.RuleCard.1'
+
+- tapOn:
+    id: 'TabBar.Transactions'
+- waitForAnimationToEnd
+
+- extendedWaitUntil:
+    visible:
+      id: 'TransactionCard.Label.E2E_Pill_Navigation'
+    timeout: 10000
+- tapOn:
+    id: 'TransactionCard.Label.E2E_Pill_Navigation'
+- waitForAnimationToEnd
+
+- extendedWaitUntil:
+    visible:
+      id: 'MatchingRules.Pill'
+    timeout: 10000
+- tapOn:
+    id: 'MatchingRules.Pill'
+- waitForAnimationToEnd
+
+- extendedWaitUntil:
+    visible:
+      id: 'RulesPage.MatchingRulesFilter'
+    timeout: 10000
+- assertVisible: '2 matching rules'
+- assertVisible:
+    id: 'RulesPage.RuleCard.0'
+- assertVisible:
+    id: 'RulesPage.RuleCard.1'
+
+- tapOn:
+    id: 'RulesPage.RuleCard.0'
+- extendedWaitUntil:
+    visible: 'Edit Rule'
+
+- tapOn:
+    id: 'PageHeader.BackButton'
+- waitForAnimationToEnd
+- assertVisible:
+    id: 'RulesPage.MatchingRulesFilter'
+
+- tapOn:
+    id: 'PageHeader.BackButton'
+- waitForAnimationToEnd
+- assertVisible: 'Edit Expense'
